# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ariadne Metadata Management API
  version: 0.1.0
  description: >
    Phase 0 contract for Ariadne backend. Covers authentication, metadata CRUD,
    lineage management (with lineage_source), bulk import/export, semantic search,
    and AI assistant endpoints. All endpoints are prefixed with `/api/v1`.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /tables/{table_id}/fields:
    get:
      tags: [fields]
      summary: List fields for a table
      parameters:
        - $ref: '#/components/parameters/table_id'
      responses:
        '200':
          description: Fields of the table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [fields]
      summary: Create field metadata for a table (table_id from path)
      parameters:
        - $ref: '#/components/parameters/table_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldCreateInTable'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        default:
          $ref: '#/components/responses/Error'
  /tables/{table_id}/fields/batch:
    post:
      tags: [fields]
      summary: Batch create fields for a table (table_id from path)
      parameters:
        - $ref: '#/components/parameters/table_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/FieldCreateInTable'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Field'
        default:
          $ref: '#/components/responses/Error'
  /fields:
    post:
      tags: [fields]
      summary: Create field metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        default:
          $ref: '#/components/responses/Error'
  /fields/{field_id}:
    get:
      tags: [fields]
      summary: Get field detail
      parameters:
        - $ref: '#/components/parameters/field_id'
      responses:
        '200':
          description: Field detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [fields]
      summary: Update field metadata
      parameters:
        - $ref: '#/components/parameters/field_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [fields]
      summary: Delete field metadata
      parameters:
        - $ref: '#/components/parameters/field_id'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /lineage/table/{table_id}/upstream:
    get:
      tags: [lineage]
      summary: Get upstream lineage graph for a table
      parameters:
        - $ref: '#/components/parameters/table_id'
        - in: query
          name: depth
          schema:
            type: integer
            minimum: 1
            maximum: 5
            default: 3
      responses:
        '200':
          description: Lineage graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageGraphResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /lineage/table/{table_id}/downstream:
    get:
      tags: [lineage]
      summary: Get downstream lineage graph for a table
      parameters:
        - $ref: '#/components/parameters/table_id'
        - in: query
          name: depth
          schema:
            type: integer
            minimum: 1
            maximum: 5
            default: 3
      responses:
        '200':
          description: Lineage graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageGraphResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /lineage/table:
    post:
      tags: [lineage]
      summary: Create table-level lineage relationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableLineageCreateRequest'
      responses:
        '201':
          description: Created lineage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageRelationship'
        '422':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /lineage/field:
    post:
      tags: [lineage]
      summary: Create field-level lineage relationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldLineageCreateRequest'
      responses:
        '201':
          description: Created lineage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageRelationship'
        '422':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /lineage/{lineage_id}:
    delete:
      tags: [lineage]
      summary: Delete a lineage relationship by Neo4j relationship id
      parameters:
        - $ref: '#/components/parameters/lineage_id'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /bulk/import:
    post:
      tags: [import_export]
      summary: Bulk import metadata (validate, preview, execute)
      description: Upload CSV/JSON/YAML/Excel with sources, tables, fields, and lineage. Choose mode to validate, preview changes, or execute transactionally.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, format]
              properties:
                file:
                  type: string
                  format: binary
                format:
                  type: string
                  enum: [csv, json, yaml, xlsx]
                mode:
                  type: string
                  enum: [validate, preview, execute]
                  default: validate
                rollback_on_error:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Bulk import result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkImportResponse'
        default:
          $ref: '#/components/responses/Error'
  /bulk/export:
    get:
      tags: [import_export]
      summary: Bulk export metadata
      parameters:
        - in: query
          name: format
          required: true
          schema:
            type: string
            enum: [csv, json, yaml, xlsx]
            default: csv
      responses:
        '200':
          description: Exported file
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: string
                format: binary
            application/x-yaml:
              schema:
                type: string
                format: binary
            application/vnd.ms-excel:
              schema:
                type: string
                format: binary
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    table_id:
      in: path
      name: table_id
      required: true
      schema:
        type: string
        format: uuid
    field_id:
      in: path
      name: field_id
      required: true
      schema:
        type: string
        format: uuid
    lineage_id:
      in: path
      name: lineage_id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Field:
      type: object
      properties:
        id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        name:
          type: string
        data_type:
          type: string
        description:
          type: string
        is_nullable:
          type: boolean
        is_primary_key:
          type: boolean
        is_foreign_key:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FieldCreateRequest:
      type: object
      required: [table_id, name, data_type]
      properties:
        table_id:
          type: string
          format: uuid
        name:
          type: string
        data_type:
          type: string
        description:
          type: string
        is_nullable:
          type: boolean
        is_primary_key:
          type: boolean
        is_foreign_key:
          type: boolean
    FieldUpdateRequest:
      type: object
      properties:
        name:
          type: string
        data_type:
          type: string
        description:
          type: string
        is_nullable:
          type: boolean
        is_primary_key:
          type: boolean
        is_foreign_key:
          type: boolean
    FieldListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Field'
...
    LineageSource:
      type: string
      enum: [inferred, manual, approved]
    LineageRelationship:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [table, field]
        source_node_id:
          type: string
          format: uuid
        target_node_id:
          type: string
          format: uuid
        transformation_type:
          type: string
          example: ETL
        transformation_logic:
          type: string
        lineage_source:
          $ref: '#/components/schemas/LineageSource'
        status:
          type: string
          enum: [pending, approved, rejected]
          default: pending
        approved_by:
          type: string
          format: email
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
        confidence:
          type: number
          format: float
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    TableLineageCreateRequest:
      type: object
      required: [source_table_id, target_table_id, lineage_source]
      properties:
        source_table_id:
          type: string
          format: uuid
        target_table_id:
          type: string
          format: uuid
        transformation_type:
          type: string
        transformation_logic:
          type: string
        lineage_source:
          $ref: '#/components/schemas/LineageSource'
        confidence:
          type: number
          minimum: 0
          maximum: 1
        metadata:
          type: object
          additionalProperties: true
    FieldLineageCreateRequest:
      type: object
      required: [source_field_id, target_field_id, lineage_source]
      properties:
        source_field_id:
          type: string
          format: uuid
        target_field_id:
          type: string
          format: uuid
        transformation_logic:
          type: string
        lineage_source:
          $ref: '#/components/schemas/LineageSource'
        confidence:
          type: number
          minimum: 0
          maximum: 1
        metadata:
          type: object
          additionalProperties: true
...
    LineageGraphNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        type:
          type: string
          enum: [table, field]
        source_id:
          type: string
          format: uuid
          nullable: true
    LineageGraphEdge:
      type: object
      properties:
        id:
          type: string
          format: uuid
        from:
          type: string
          format: uuid
        to:
          type: string
          format: uuid
        lineage_source:
          $ref: '#/components/schemas/LineageSource'
        confidence:
          type: number
          nullable: true
        metadata:
          type: object
          additionalProperties: true
    LineageGraphResponse:
      type: object
      properties:
        root_id:
          type: string
          format: uuid
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/LineageGraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/LineageGraphEdge'
...
    BulkImportResponse:
      type: object
      properties:
        mode:
          type: string
          enum: [validate, preview, execute]
        success:
          type: boolean
        summary:
          type: object
          properties:
            total_rows:
              type: integer
            created:
              type: integer
            updated:
              type: integer
            skipped:
              type: integer
        errors:
          type: array
          description: Row-level errors for validation/preview/execute
          items:
            type: object
            properties:
              row:
                type: integer
              entity:
                type: string
                example: table
              field:
                type: string
              message:
                type: string
              code:
                type: string
                example: SCHEMA_MISMATCH
        preview:
          type: object
          description: Present when mode=preview
          properties:
            to_create:
              type: array
              items:
                type: object
            to_update:
              type: array
              items:
                type: object
        transaction_id:
          type: string
          description: Identifier for executed import

      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        progress:
          type: number
          minimum: 0
          maximum: 1
        stats:
          type: object
          properties:
            processed_rows:
              type: integer
            succeeded_rows:
              type: integer
            failed_rows:
              type: integer
        error_report_url:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    FieldCreateInTable:
      type: object
      required: [name, data_type]
      properties:
        name:
          type: string
        data_type:
          type: string
        description:
          type: string
        is_nullable:
          type: boolean
        is_primary_key:
          type: boolean
        is_foreign_key:
          type: boolean